Explicity check for window focus before refocusing (#265)
Under some certain circumstances in a desktop environment running on a Xorg server, the Treesheets window catches all input events - even those that are not intended for Treesheets. This leads Treesheets to set the focus on its current tab and thus on its window each time an input event is detected - even if the window had no focus. Explicitly check whether the Treesheets window has focus and only if it has focus, refocus on the Treesheets current tab and window.